requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
click==8.1.8
dnspython==2.7.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
Jinja2==3.1.5
jose==1.0.0
MarkupSafe==3.0.2
motor==3.7.0
numpy==2.2.3
opencv-python==4.11.0.86
passlib==1.7.4
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
python-dotenv==1.0.1
python-multipart==0.0.20
qrcode==8.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
